fix(event-comments): prevent pagination offset overflow

+++ backend/src/main/java/nl/hackyourfuture/project/backend/event/comment/repository/EventCommentRepository.java
@@ -57,7 +57,7 @@ public class EventCommentRepository {
     public List<EventComment> findCommentsByEventId(
             UUID eventId,
             int limit,
-            int offset
+            long offset
     ) {
         String sql = selectCommentWithUserFrom("event_comments") + """
                 WHERE c.event_id = :eventId
